% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_starts.R
\docType{data}
\name{method_Menzel}
\alias{method_Menzel}
\title{Vegetation Start Method "Menzel"}
\description{
The method "Menzel" implements the algorithm described in Menzel (1997).
The method is parameterized for 10 common tree species ( \emph{Larix decidua},
\emph{Picea abies (frueh)}, \emph{Picea abies (spaet)}, \emph{Picea abies (noerdl.)},
\emph{Picea omorika}, \emph{Pinus sylvestris}, \emph{Betula pubescens}, \emph{Quercus robur},
\emph{Quercus petraea}, \emph{Fagus sylvatica}). It needs previous year's chill days.
}
\section{Calculations}{


\if{html}{\out{<div class="sourceCode">}}\preformatted{ - counting chill days in Nov & Dec of previous year if previous year available else estimate previous year's chill days
 - count chill days in current year
 - determine critical temperature per days using regression equation fitted by Menzel 1997
 - calculate HeatSum starting from Feb
 - vegetation start if HeatSum >= criticalTemp
}\if{html}{\out{</div>}}
}

\references{
Menzel, A. (1997) Phänologie von Waldbäumen unter sich ändernden Klimabedingungen - Auswertung der Beobachtungen in den Internationalen Phänologischen Gärten und Möglichkeiten der Modellierung von Phänodaten. Forstliche Forschungsberichte München.
}
\keyword{internal}
