% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_ends.R
\name{method_NuskeAlbert}
\alias{method_NuskeAlbert}
\title{Vegetation End Method "NuskeAlbert"}
\description{
The method \code{NuskeAlbert} provides a very simple method which is inspired by
standard climatological procedures but employs a 7 day moving average and
a 5 °C threshold (cf. Walther and Linderholm 2006).
}
\section{Calculations}{

2 criteria: temperature & short day criterion
\subsection{Temperature criterion}{

7 day moving average of daily mean temperatures
end if 5 consecutive days under under 5° C
start the search at 1st of July / vegetation start
}

\subsection{Short day criterion}{

last day of the vegetation period is 5th of October
}
}

\references{
Walther, A. and Linderholm, H. W. (2006)
A comparison of growing season indices for the Greater Baltic Area.
\emph{International Journal of Biometeorology}, \bold{51}(2), 107--118.
\doi{10.1007/s00484-006-0048-5}.
}
\keyword{internal}
