% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/veg_ends.R
\name{method_vonWilpert}
\alias{method_vonWilpert}
\title{Vegetation End Method "vonWilpert"}
\description{
The method \code{vonWilpert} is based on von Wilpert (1990). It was originally
developed for "Picea abies" in the Black Forest but is commonly used for all
tree species throughout Germany. As usual, the rules regarding the soilmatrix
are neglected in this implementation.
}
\section{Calculations}{


Orthodox are 3 criteria: \strong{short day}, \strong{temperature} and \strong{drought criterion}
we consider -as usual- only short day and temperature citerion
\subsection{Temperature criterion:}{
\itemize{
\item 7 day moving average of daily mean temperatures
at least 5 consecutive days under 10°C
\item if afterwards more than 5 consecutive days 7 day moving average over 10°C
vegetation period gets restarted
}
}

\subsection{Short day criterion}{
\itemize{
\item last day of the vegetation period is DOY 279 (5th of October in leap years)
}
}
}

\references{
von Wilpert, K. (1990)
Die Jahrringstruktur von Fichten in Abhängigkeit vom Bodenwasserhaushalt
auf Pseudogley und Parabraunerde: Ein Methodenkonzept zur Erfassung
standortsspezifischer Wasserstreßdispostion.
\emph{Freiburger Bodenkundliche Abhandlungen}. Pages 106--108.
}
\keyword{internal}
