% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dwd.R
\name{read.DWDdata}
\alias{read.DWDdata}
\title{Fetch meteo data of DWD weather stations}
\usage{
read.DWDdata(id, type = "climate", period = "recent",
  resolution = "daily", file = NULL, destdir = NULL, quiet = FALSE)
}
\arguments{
\item{id}{integer. A valid station id (cf. \link{read.DWDstations}).}

\item{type}{string. Stations can be of type \samp{"climate"} offering a range
of meteorological parameters or of type \samp{"precipitation"} solely
recording rainfall. There are about 400 \samp{climate} and 2000
\samp{precipitation} stations.}

\item{period}{string specifying a time span. The last 1.5 years are
called \samp{"recent"}. Data from the individual beginning of measurements
up to a year ago are labelled \samp{"historical"}.
\samp{"recent"} data have not yet gone through the full quality control
procedure. \samp{"historical"} data have completed the operational quality
control. But be aware there still might be gaps and inhomogeneities.}

\item{resolution}{string. Temporal resolution of the data can be
\samp{"daily"} or \samp{"monthly"}. \samp{"daily"} being the common
resolution.}

\item{file}{a string specifying the location of a dataset. \samp{"file"}
may point to a file on a FTP-server or on the local file system.
If the file lives on a FTP-server the string must start with
\code{'ftp://'}.}

\item{destdir}{directory (string) where intermediate data (downloaded *.zip
file) are stored. If it is NULL (the default) a subdirectory
\samp{downloaded_packages} of the session temporary directory will be used
(and the files will be deleted at the end of the session).
It might be advisable to keep the .zip file since it contains the
stations documentation and metadata (eg. station shift, used devices,
downtimes etc.).}

\item{quiet}{If TRUE, suppress status messages (if any), and the progress bar.}
}
\value{
A data.frame with the observed weather data. Be aware there might be
  gaps and inhomogeneities! \samp{colnames} contains the original header and
  hence German terms. Translation of usual column names:
  \tabular{ll}{
   \strong{original name} \tab \strong{translation}\cr
   STATIONS_ID              \tab id\cr
   MESS_DATUM               \tab date\cr
   QUALITAETS_NIVEAU        \tab quality level\cr
   LUFTTEMPERATUR           \tab average temperature\cr
   DAMPFDRUCK               \tab vapor pressure\cr
   BEDECKUNGSGRAD           \tab cloud cover\cr
   LUFTDRUCK_STATIONSHOEHE  \tab air pressure\cr
   REL_FEUCHTE              \tab humidity\cr
   WINDGESCHWINDIGKEIT      \tab wind speed\cr
   LUFTTEMPERATUR_MAXIMUM   \tab max temperature\cr
   LUFTTEMPERATUR_MINIMUM   \tab min temperature\cr
   LUFTTEMP_AM_ERDB_MINIMUM \tab min temperature at ground level\cr
   WINDSPITZE_MAXIMUM       \tab max wind speed\cr
   NIEDERSCHLAGSHOEHE       \tab precipitation\cr
   NIEDERSCHLAGSHOEHE_IND   \tab type of precipitation\cr
   SONNENSCHEINDAUER        \tab sunshine duration\cr
   SCHNEEHOEHE              \tab snow height
  }
}
\description{
Fetch observed meteorological data of German weather stations from freely
accessible part of the Climate Data Center of the German Weather Service
(Deutscher Wetterdienst, DWD).
}
\details{
An introduction to the data available at the Climate Data Center can be
found in the German
\href{../doc/Liesmich_intro_CDC_ftp.pdf}{Liesmich_intro_CDC_ftp}
or the translated
\href{../doc/Readme_intro_CDC_ftp.pdf}{Readme_intro_CDC_ftp}.

The freely accessible part of the Climate Data Center of the German Weather
Service (Deutscher Wetterdienst, DWD) is part of the DWD's mandate for
basic supply of information (termed "Grundversorgung"). These services may
be used without any restrictions (no fees will be charged and in general
there are no restrictions for the use the data), provided that the source is
indicated as laid down in the
\href{http://www.gesetze-im-internet.de/geonutzv/BJNR054700013.html}{"GeoNutzV"}
ordinance. The source reference shall meet the following rules:
\itemize{
  \item Where data are used without modification, the source reference shall
    read "Source: Deutscher Wetterdienst" or just consist of the DWD logo.
  \item If the data are modifyed, the source reference shall specify as
    precisely as possible the extent of such, e.g. "Based on data from
    Deutscher Wetterdienst, figures rounded".
}
A more detailed description of the rules can be found in the official German
\href{../doc/Nutzungsbedingungen_German.pdf}{Nutzungsbedingungen}
or the translated \href{../doc/Terms_of_use.pdf}{Terms of use}.
}
\examples{
\dontrun{
# fetch last 500 days worth of data from station Göttingen
clim <- read.DWDdata(id=1691)

# save data & metadata (documentation about devices, downtimes etc.)
clim <- read.DWDdata(id=1691, destdir='.')

# find and download historical data from the Brocken
stat.hist <- read.DWDstations(period='historical')
brocken.id <- stat.hist[grep("^Brock", stat.hist$name), ]$id
clim.brocken <- read.DWDdata(id=brocken.id, period='historical')

# get recent meteorological data fom a randomly selected station
require('RCurl')
dwdURL <- paste0("ftp://ftp-cdc.dwd.de/pub/CDC/observations_germany/",
                 "climate/daily/kl/recent/")
filenames <- getURL(dwdURL, ftp.use.epsv=FALSE, dirlistonly=TRUE)
filenames <- strsplit(filenames, '\\r*\\n')[[1]]
stations <- grep("Beschreibung_Stationen.txt$", filenames,
                 invert=TRUE, value=TRUE)
rnd.station <- sample(stations, 1)
rnd.clim <- read.DWDdata(file=paste0(dwdURL, rnd.station))
}

}
\references{
Freely accessible DWD data are available via the
  \href{http://www.dwd.de/cdc}{Climate Data Center}.
}
