% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dwd.R
\name{read.DWDstations}
\alias{read.DWDstations}
\title{Fetch information about DWD weather stations}
\usage{
read.DWDstations(type = "climate", period = "recent",
  resolution = "daily")
}
\arguments{
\item{type}{string. Stations can be of type \samp{"climate"} offering a range
of meteorological parameters or of type \samp{"precipitation"} solely
recording rainfall.}

\item{period}{string specifying a time span. The last 1.5 years are
called \samp{"recent"}. Data from the individual beginning of measurements
up to a year ago are labelled \samp{"historical"}.}

\item{resolution}{string. Temporal resolution of the data can be
\samp{"daily"} or \samp{"monthly"}. \samp{"daily"} being the common
resolution.}
}
\value{
A data.frame with all advertised weather stations. The data.frame
contains the variables id, from, to, elev, lat, long, name, state.
}
\description{
Get a listing of advertised stations of the German Weather Service (Deutscher
Wetterdienst, DWD). The list differs for station type, time period and
resolution. The list of advertised stations is usually not identical with
the currently available stations.
}
\details{
The freely accessible part of the Climate Data Center of the German Weather
Service (Deutscher Wetterdienst, DWD) is part of the DWD's mandate for
basic supply of information (termed "Grundversorgung"). These services may
be used without any restrictions (no fees will be charged and in general
there are no restrictions for the use the data), provided that the source is
indicated as laid down in the
\href{http://www.gesetze-im-internet.de/geonutzv/BJNR054700013.html}{"GeoNutzV"}
ordinance. The source reference shall meet the following rules:
\itemize{
  \item Where data are used without modification, the source reference shall
    read "Source: Deutscher Wetterdienst" or just consist of the DWD logo.
  \item If the data are modifyed, the source reference shall specify as
    precisely as possible the extent of such, e.g. "Based on data from
    Deutscher Wetterdienst, figures rounded".
}
A more detailed description of the rules can be found in the official German
\href{../doc/Nutzungsbedingungen_German.pdf}{Nutzungsbedingungen}
or the translated \href{../doc/Terms_of_use.pdf}{Terms of use}.
}
\examples{
\dontrun{
# stations with daily climate data from the last year
stations <- read.DWDstations()

# precipitation stations with monthly histrical data
precip <- read.DWDstations(type='precip', period='historical',
                           resolution='monthly')

# list stations updated within last week
stat.daily <- read.DWDstations(period='recent')
stat.daily.recent <- stat.daily[stat.daily$to > (Sys.Date() - 7), ]
}

}
\references{
Freely accessible DWD data are available via the
  \href{http://www.dwd.de/cdc}{Climate Data Center}
}
