% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vegperiod-package.R
\docType{package}
\name{vegperiod-package}
\alias{vegperiod-package}
\title{The vegperiod Package}
\description{
Determine Vegetation Period of Forest Trees.
}
\details{
\tabular{ll}{
Package: \tab vegperiod\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2015-04-06\cr
Depends: \tab R (>= 3.1.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://computerfoerster.de/r-pkgs\cr
}

Calculate start and end date of vegetation periods based on daily average air temperature and the day of the year (DOY).
}
\author{
Robert Nuske
}
\keyword{package}

