% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vegperiod.R
\name{vegperiod}
\alias{vegperiod}
\title{Determine vegetation period.}
\usage{
vegperiod(dates, Tavg, species, method.start, method.end, first.avg = 0,
  temp.sum = FALSE)
}
\arguments{
\item{dates}{vector calender dates as object of class \code{Date}.
Same length as Tavg.}

\item{Tavg}{vector of daily average air temperatures. Same length as dates.}

\item{species}{name of tree species. Needed for start of vegetation
calculation according to Menzel (1997). Must be one of "Larix decidua",
"Picea abies (frueh)", "Picea abies (spaet)",
"Picea abies (noerdl.)", "Picea omorika", "Pinus sylvestris",
"Betula pubescens", "Quercus robur", "Quercus petraea",
"Fagus sylvatica".}

\item{method.start}{name of method used to calculate the start of vegetation
period. At the moment one of "StdClimatology", "Menzel".}

\item{method.end}{name method used to calculate the end of vegetaion period.
At the moment one of "vonWilpert", "LWF-BROOK90", "NuskeAlbert" and
"StdClimatology".}

\item{first.avg}{integer. Guesstimate number of previous years'
\code{cold days} (needed by Menzel, 1997) from the \code{first.avg}
first years. If \code{first.avg=0}, first year is only used for
calculating \code{cold days} and discarded afterwards ->
losing one year from the provided data.}

\item{temp.sum}{boolean. Return the sum of day degrees for the vegetation
  period.}
}
\value{
Returns a data.frame with year and DOY of start and end day of
  vegetation period. If \code{temp.sum=TRUE}, the data.frame contains an
  additional column with the sum of day degree within the vegetation period.
}
\description{
\code{vegperiod} calculates start and end day of vegetation period based on
daily average air temperature and the day of the year (DOY).
}
\details{
Implemented are at moment the simple standard procedure used in climatology
(FORMAYER et al., 2007), vegetation start according to Menzel (1997) and
vegetation end accodring to von Wilpert (1990), a reimplementation of
vonWilpert found in LWF-BROOK90 and a very simple method by Nuske & Albert.

The LWF-BROOK90 reimplementation was tested against the original code using
data from a range of different climate stations.
}
\examples{
data(goe)
vegperiod(dates=goe$date, Tavg=goe$t, species="Picea abies (frueh)",
  first.avg=5, method.start="Menzel", method.end="vonWilpert")
}

