% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vegperiod-package.R
\docType{package}
\name{vegperiod-package}
\alias{vegperiod-package}
\title{Determine Vegetation Period of Forest Trees}
\description{
Determine Vegetation Period of Forest Trees.
}
\details{
\tabular{ll}{
Package: \tab vegperiod\cr
Type: \tab Package\cr
Version: \tab 0.2.1\cr
Date: \tab 2015-06-11\cr
Depends: \tab R (>= 3.1.0)\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab http://computerfoerster.de/r-pkgs\cr
}

Calculate start and end date of vegetation period based on daily average air temperature and the day of the year (DOY). Functions for summing day degrees within vegetation period and reading freely available meteo data from DWD's ftp-Server are provided on top.
}
\author{
Robert Nuske
}
\keyword{package}

